change1

+++ window_scene/session/host/src/scene_session.cpp
@@ -1354,7 +1354,7 @@ WSError SceneSession::UpdateRect(const WSRect& rect, SizeChangeReason reason,
             session->NotifyClientToUpdateRect(updateReason, rsTransaction);
         }
         session->dirtyFlags_ |= static_cast<uint32_t>(SessionUIDirtyFlag::RECT);
-        session->AddPropertyDirtyFlag(SessionPropertyFlag::RECT);
+        session->AddPropertyDirtyFlag(static_cast<uint32_t>(SessionPropertyFlag::RECT));
         TLOGNI(WmsLogTag::WMS_LAYOUT, "%{public}s: id:%{public}d, reason:%{public}d %{public}s, "
             "rect:%{public}s, clientRect:%{public}s",
             where, session->GetPersistentId(), session->reason_, updateReason.c_str(),
